backend/requirements.txt
flask
flask-cors
ultralytics
opencv-python
pillow
pyyaml
numpy
